requirements.txt
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
Flask~=3.1.0
youtube_transcript_api==0.6.3
Flask-Cors==5.0.0
google==3.0.0
google.genai==1.7.0